# api/openapi.json
{
  "components": {
    "schemas": {
      "AuthResponse": {
        "additionalProperties": false,
        "properties": {
          "user": {
            "$ref": "#/components/schemas/MeResponse"
          }
        },
        "required": [
          "user"
        ],
        "title": "AuthResponse",
        "type": "object"
      },
...
      "ForgotPasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          }
        },
        "required": [
          "email"
        ],
        "title": "ForgotPasswordRequest",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "LoginRequest": {
        "additionalProperties": false,
        "properties": {
          "email": {
            "format": "email",
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 1,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password"
        ],
        "title": "LoginRequest",
        "type": "object"
      },
      "MeResponse": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "format": "email",
            "title": "Email",
            "type": "string"
          },
          "emailVerified": {
            "title": "Emailverified",
            "type": "boolean"
          },
          "preferences": {
            "$ref": "#/components/schemas/PreferencesResponse"
          }
        },
        "required": [
          "email",
          "displayName",
          "emailVerified",
          "preferences"
        ],
        "title": "MeResponse",
        "type": "object"
      },
...
      "PreferencesResponse": {
        "additionalProperties": false,
        "properties": {
          "language": {
            "title": "Language",
            "type": "string"
          },
          "theme": {
            "title": "Theme",
            "type": "string"
          },
          "version": {
            "minimum": 1.0,
            "title": "Version",
            "type": "integer"
          }
        },
        "required": [
          "theme",
          "language",
          "version"
        ],
        "title": "PreferencesResponse",
        "type": "object"
      },
      "RegisterRequest": {
        "additionalProperties": false,
        "properties": {
          "displayName": {
            "maxLength": 120,
            "minLength": 1,
            "title": "Displayname",
            "type": "string"
          },
          "email": {
            "format": "email",
            "maxLength": 320,
            "title": "Email",
            "type": "string"
          },
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "email",
          "password",
          "displayName"
        ],
        "title": "RegisterRequest",
        "type": "object"
      },
      "ResetPasswordRequest": {
        "additionalProperties": false,
        "properties": {
          "password": {
            "format": "password",
            "maxLength": 128,
            "minLength": 12,
            "title": "Password",
            "type": "string",
            "writeOnly": true
          },
          "token": {
            "format": "password",
            "maxLength": 512,
            "minLength": 32,
            "title": "Token",
            "type": "string",
            "writeOnly": true
          }
        },
        "required": [
          "token",
          "password"
        ],
        "title": "ResetPasswordRequest",
        "type": "object"
      },
      "StatusResponse": {
        "additionalProperties": false,
        "properties": {
          "status": {
            "const": "ok",
            "default": "ok",
            "title": "Status",
            "type": "string"
          }
        },
        "title": "StatusResponse",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "SessionCookie": {
        "description": "Secure, HttpOnly current-account session cookie.",
        "in": "cookie",
        "name": "__Host-qg_session",
        "type": "apiKey"
      },
      "SessionCsrf": {
        "description": "Readable session-bound CSRF cookie value echoed by the same-origin client.",
        "in": "header",
        "name": "X-CSRF-Token",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "title": "QuantGym API",
    "version": "2.0.0-preview.1"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v2/auth/google/callback": {
      "get": {
        "operationId": "completeGoogleAuthentication",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 4096,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Code"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "State"
            }
          },
          {
            "in": "query",
            "name": "error",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 256,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Error"
            }
          }
        ],
        "responses": {
          "303": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Google authentication and rotate the session",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v2/auth/google/start": {
      "get": {
        "operationId": "startGoogleAuthentication",
        "parameters": [
          {
            "in": "query",
            "name": "redirectPath",
            "required": false,
            "schema": {
              "default": "/",
              "maxLength": 512,
              "minLength": 1,
              "title": "Redirectpath",
              "type": "string"
            }
          }
        ],
        "responses": {
          "302": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Start Google authorization-code authentication",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v2/auth/login": {
      "post": {
        "operationId": "loginAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Authenticate a local account and rotate the session",
        "tags": [
          "authentication"
        ]
      }
    },
...
    "/api/v2/auth/password/forgot": {
      "post": {
        "operationId": "requestPasswordReset",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ForgotPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Request an enumeration-resistant password reset",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v2/auth/password/reset": {
      "post": {
        "operationId": "resetPassword",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResetPasswordRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/StatusResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Consume a one-time reset challenge and replace the password",
        "tags": [
          "authentication"
        ]
      }
    },
    "/api/v2/auth/register": {
      "post": {
        "operationId": "registerAccount",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AuthResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Create a local account and rotate into a session",
        "tags": [
          "authentication"
        ]
      }
    },
...
  }
}